refactor: Move all fetcher helper functions to lib
