+++ etc/fetchers/fetch-wbw.ts
@@ -1,20 +1,11 @@
+import { toSecond } from "../../lib/string-utils";
+
 import { fetchDocs } from "./fetch-docs";
 import { fetchFaqSheets } from "./fetch-faq-sheets";
 import { fetchSheets } from "./fetch-sheets";
 
 import chalk from "chalk";
 import ora from "ora";
-
-/**
- * Get current time in seconds
- *
- * @param {[number, number]} hrtime High resolution time, should be
- * `process.hrtime()`
- * @returns {string} time string, in seconds.
- */
-function toSecond(hrtime: [number, number]): string {
-  return (hrtime[0] + hrtime[1] / 1e9).toFixed(3);
-}
 
 (function fetchWbw() {
   const start = process.hrtime();
